Add unit test + fix PR feedback
