+++ Client/Frontend/Browser/Tabs/Legacy/LegacyTabTrayViewModel.swift
@@ -6,45 +6,6 @@ import Common
 import Shared
 import Storage
 
-enum TabTrayPanelType: Int, CaseIterable {
-    case tabs
-    case privateTabs
-    case syncedTabs
-
-    var navTitle: String {
-        switch self {
-        case .tabs:
-            return .TabTrayV2Title
-        case .privateTabs:
-            return .TabTrayPrivateBrowsingTitle
-        case .syncedTabs:
-            return .AppMenu.AppMenuSyncedTabsTitleString
-        }
-    }
-
-    var label: String {
-        switch self {
-        case .tabs:
-            return String.TabTraySegmentedControlTitlesTabs
-        case .privateTabs:
-            return String.TabTraySegmentedControlTitlesPrivateTabs
-        case .syncedTabs:
-            return String.TabTraySegmentedControlTitlesSyncedTabs
-        }
-    }
-
-    var image: UIImage? {
-        switch self {
-        case .tabs:
-            return UIImage(named: ImageIdentifiers.navTabCounter)
-        case .privateTabs:
-            return UIImage(named: ImageIdentifiers.privateMaskSmall)
-        case .syncedTabs:
-            return UIImage(named: ImageIdentifiers.syncedDevicesIcon)
-        }
-    }
-}
-
 class LegacyTabTrayViewModel {
     enum Layout: Equatable {
         case regular // iPad
